Move codegen package from cmd/ to internal/
