+++ internal/codegen/jsonschema.go
