+++ internal/codegen/kindregistry.go
