+++ internal/codegen/kindsyscomposable.go
