+++ internal/codegen/kindsyscore.go
